feat: damage deposits repository and validations

+++ app/Http/Controllers/DamageDepositsController.php
@@ -4,15 +4,14 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Helpers\LanguageHelper;
-use App\Repositories\DemageDepositsRepositoryInterface;
-use App\Models\{ DamageDeposit, DamageDepositsTranslation };
+use App\Repositories\DamageDepositsRepositoryInterface;
+use App\Models\DamageDeposit;
 
 class DamageDepositsController extends Controller
 {
-
     private $repository;
 
-    public function __construct(DemageDepositsRepositoryInterface $repository) 
+    public function __construct(DamageDepositsRepositoryInterface $repository) 
     {
         $this->repository = $repository;
     }
@@ -23,10 +22,10 @@ class DamageDepositsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index(Request $request)
-    {   
+    {
         $search = trim($request->s);
         $damage_deposits = $this->repository->all($search);
-        
+
         return view('damage-deposits.index')
             ->with('damage_deposits', $damage_deposits)
             ->with('search', $search);
@@ -50,12 +49,11 @@ class DamageDepositsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
-    {
-
+    { 
         $damage_deposit = $this->repository->create($request);
+        $request->session()->flash('success', __('Record created successfully'));
         return redirect(route('damage-deposits.edit', [$damage_deposit->id]));
-    }
-
+}
     /**
      * Display the specified resource.
      *
@@ -63,8 +61,8 @@ class DamageDepositsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function show(DamageDeposit $damage_deposit)
-    {   
-        $damage_deposit = $this->repository->find($damage_deposit);        
+    {
+        $damage_deposit = $this->repository->find($damage_deposit);
         return view('damage-deposits.show')->with('damage_deposit', $damage_deposit);
     }
 
@@ -90,6 +88,7 @@ class DamageDepositsController extends Controller
     public function update(Request $request, $id)
     {
         $this->repository->update($request, $id);
+        $request->session()->flash('success', __('Record updated successfully'));
         return redirect(route('damage-deposits.edit', [$id]));
     }
 
@@ -99,9 +98,15 @@ class DamageDepositsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $this->repository->delete($id);
-        return redirect(route('damage-deposits'));
+        if ( $this->repository->canDelete($id) ) {
+            $this->repository->delete($id);
+            $request->session()->flash('success', __('Record deleted successfully'));
+            return redirect(route('damage-deposits'));
+        }
+
+        $request->session()->flash('error', __("This record can't be deleted"));
+        return redirect()->back();
     }
 }
